lowercase routes; reorganize routes and pages for CRU capabilities
