+++ frontend/src/App.jsx
@@ -1,24 +1,26 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import ProfilePage from './pages/ProfilePage';
+import ProfilePage from "./pages/ProfilePage";
 import HomePage from "./pages/Home";
 import "./App.css";
 import ExperienceSearch from "./pages/ExperienceSearch";
-import TripPage from "./pages/TripPage";
+import ViewTrip from "./pages/ViewTrip";
 import LoginForm from "./pages/Login";
-import AddTrip from './pages/AddTrip';
+import AddTrip from "./pages/AddTrip";
 
 const App = () => {
   return (
     <div>
       <Router>
         <Routes>
-          <Route path="/ProfilePage" element={<ProfilePage />} />
-          <Route path="/ExperienceSearch" element={<ExperienceSearch />} />
+          <Route path="/" element={<HomePage />} />
+          <Route path="/login" element={<LoginForm />} />
+          <Route path="/profilepage" element={<ProfilePage />} />
+          <Route path="/experiencesearch" element={<ExperienceSearch />} />
           <Route
-            path="/TripPage"
+            path="/trip"
             element={
-              <TripPage
+              <ViewTrip
                 trip={{
                   id: 1,
                   name: "First Test Trip",
@@ -30,9 +32,7 @@ const App = () => {
               />
             }
           />
-          <Route path='/AddTrip' element={<AddTrip />}/>
-          <Route path="/login" element={<LoginForm />} />
-          <Route path="/" element={<HomePage />} />
+          <Route path="/addtrip" element={<AddTrip />} />
         </Routes>
       </Router>
     </div>
