+++ frontend/src/pages/Home.jsx
@@ -6,19 +6,19 @@ function HomePage() {
       <h1>TripXP HomePage</h1>
       <ul>
         <li>
-          <a href="/ProfilePage">Profile Page</a>
-        </li>
-        <li>
-          <a href="/ExperienceSearch">Experience Search</a>
-        </li>
-        <li>
-          <a href="/TripPage">Trip Page</a>
+          <a href="/profilepage">Profile Page</a>
         </li>
         <li>
           <a href="/login">Login</a>
         </li>
         <li>
-          <a href="/AddTrip">Add Trip</a>
+          <a href="/experiencesearch">Experience Search</a>
+        </li>
+        <li>
+          <a href="/trip">View Trip</a>
+        </li>
+        <li>
+          <a href="/addtrip">Add Trip</a>
         </li>
       </ul>
     </div>
