+++ frontend/src/pages/ViewTrip.jsx
